requirements.txt
BeautifulSoup==3.2.1
bootstrap-admin==0.3.6
chardet==2.3.0
Django==1.8.14
django-simple-history==1.8.1
html5lib==1.0b3
Pillow==2.6.1
pisa==3.0.33
py2-ipaddress==3.4
pyPdf==1.13
python-ldap==2.4.19
reportlab==2.5
six==1.8.0
wheel==0.24.0
xhtml2pdf==0.0.6
xlwt==0.7.4
django-autocomplete-light==3.1.5
django-modalview==0.1.5
django-import-export==0.4.5
django-solo==1.1.2
django-mail-templated==2.6.2
django-widget-tweaks==1.4.1
django-filter==0.14.0
django-pagination==1.0.7











